+++ lib/blackjack.rb
@@ -1,3 +1,6 @@
+require_relative './card.rb'
+require_relative './hand.rb'
+
 class Blackjack
 
   attr_accessor :player_hand, :dealer_hand, :player_bankroll, :player_bet
@@ -117,98 +120,5 @@ class Blackjack
     else
       :draw
     end
-  end
-end
-
-class Hand
-  attr_accessor :cards, :name
-
-  def self.deserialize(str)
-    hand = new
-    str.chars.each_slice(2) do |card|
-      suit = Blackjack::SUITS.invert[card[1]]
-      rank = Blackjack::RANKS.invert[card[0]]
-      hand.cards << Card.new(rank: rank,suit: suit)
-    end
-    hand
-  end
-
-  def initialize(cards = [])
-    @cards = cards
-  end
-
-  def aces
-    @cards.select(&:ace?)
-  end
-
-  def best_value
-    total = @cards.inject(0) { |sum, card| sum += card.value }
-
-    aces.each do |ace|
-      break if total <= 21
-      total -= 10
-    end
-
-    total
-  end
-
-  def hard_value
-    total = @cards.inject(0) { |sum, card| sum += card.value }
-
-    aces.each do |ace|
-      total -= 10
-    end
-
-    total
-  end
-
-  def serialize
-    @cards.map(&:to_s).join
-  end
-
-end
-
-class Card
-  attr_accessor :rank, :suit
-
-  UNICODE = {
-    "Spades" => "\u2660",
-    "Diamonds" => "\u2666",
-    "Hearts" => "\u2665",
-    "Clubs" => "\u2663",
-  }
-
-  def initialize(rank:, suit:)
-    @rank = rank
-    @suit = suit
-  end
-
-  def to_s
-    "#{Blackjack::RANKS[@rank]}#{Blackjack::SUITS[@suit]}"
-  end
-
-  def ace?
-    @rank == "A"
-  end
-
-  def unicode_suit
-    UNICODE[@suit]
-  end
-
-  def color
-    case @suit
-    when "Diamonds"
-      "red"
-    when "Hearts"
-      "red"
-    when "Clubs"
-      "black"
-    when "Spades"
-      "black"
-    end
-  end
-
-  def value
-    Blackjack::VALUES[@rank]
   end
 end
